+++ CHANGELOG.md
@@ -12,6 +12,9 @@ If you like this project and find it useful, please consider giving it a star on
 
 ### Added
 
+- [shelly]: The Input component is not added when is not enabled in the device web page.
+- [shelly]: Input component configured like Button are now supported as momentary switch if exposeInputEvent=momentary.
+
 ### Verified
 
 - [shelly]: Verified shellyplus010v with firmware v. 1.4.2 and added Jest test.